class to functional
